[SofaQtQuickGUI] Rename directories (windows and helper) so they starts with a uppecase
