+++ applications-dev/plugins/SofaQtQuickGUI/src/SofaQtQuickGUI/Helper/QMathExtensions.h
