+++ applications-dev/plugins/SofaQtQuickGUI/src/SofaQtQuickGUI/Windows/CameraView.h
