+++ applications-dev/plugins/SofaQtQuickGUI/src/SofaQtQuickGUI/Windows/EditView.h
